Move TestEmbedder into test code
