+++ citatio/tests/test_models.py
@@ -1,9 +1,8 @@
 import json
 
-import numpy as np
 import pytest
 
-from citatio.models import Block, ControlFlowGraph, TestEmbedder
+from citatio.models import Block, ControlFlowGraph
 
 
 def test_empty():
@@ -49,13 +48,3 @@ def test_back_and_forth():
     )
 
     assert cfg == ControlFlowGraph.from_str(json.dumps(list(cfg)))
-
-
-def test_test_embedder():
-    data = '[[0, ["add x1,x1", "ret"]], [12, ["ret"]], [34, ["b 0"]]]'
-    embedder = TestEmbedder()
-
-    assert len(embedder.encode(data)) == 768
-    assert np.allclose(embedder.encode(data), embedder.encode(data, architecture='mips'))
-    assert np.allclose(TestEmbedder().encode(data), embedder.encode(data))
-    assert not np.allclose(embedder.encode('other data'), embedder.encode(data))
